[TS-33] Refactor: deleteReasonModal, leaveReasonModal change to common component
