+++ src/constants/modalConstants.ts
@@ -12,5 +12,5 @@ export const modalType = {
 	HABIT_MODIFY_MODAL: "HABIT_MODIFY_MODAL",
 	HABIT_CHECK_MODAL: "HABIT_CHECK_MODAL",
 	STOP_HABIT: "STOP_HABIT",
-	DELETE_REASON: "DELETE_REASON",
+	SELECT_REASON: "SELECT_REASON",
 };
