+++ src/pages/HabitDeletePage.tsx
@@ -2,7 +2,6 @@ import Header from "@/components/common/Header/Header";
 import SelectReason from "@/components/common/SelectReason/SelectReason";
 
 import { deleteReasonData } from "@/constants/homeConstants";
-import { modalType } from "@/constants/modalConstants";
 
 const HabitDeletePage = () => {
 	return (
@@ -13,7 +12,6 @@ const HabitDeletePage = () => {
 			<SelectReason
 				title="어떤 이유로 그만두시나요?"
 				reasonDefaultText="그만두는 이유를 선택해 주세요"
-				modalType={modalType.DELETE_REASON}
 				selectData={deleteReasonData}
 				footerBtnText="그만두기"
 			/>
