+++ src/pages/WithdrawalPage.tsx
@@ -1,7 +1,6 @@
 import Header from "@/components/common/Header/Header";
 import SelectReason from "@/components/common/SelectReason/SelectReason";
 
-import { modalType } from "@/constants/modalConstants";
 import { leaveReasonData } from "@/constants/myPageConstants";
 
 const WithdrawalPage = () => {
@@ -13,7 +12,6 @@ const WithdrawalPage = () => {
 			<SelectReason
 				title="어떤 이유로 탈퇴하시나요?"
 				reasonDefaultText="탈퇴 이유를 선택해 주세요"
-				modalType={modalType.LEAVE_REASON}
 				selectData={leaveReasonData}
 				footerBtnText="작별하기"
 			/>
